+++ src/plugins/ncr/sagas/changeNodeFlow.js
@@ -14,6 +14,9 @@ import waitForFlow from './waitForFlow';
 export function* successFlow(successMessage, onAfterSuccessFlow = noop) {
   yield call(onAfterSuccessFlow);
   yield call([toast, 'close']);
+  if (!successMessage) {
+    return;
+  }
   yield put(sendAlert({
     type: 'success',
     isDismissible: true,