+++ src/plugins/ncr/sagas/updateNodeFlow.js
@@ -4,7 +4,9 @@ import clearResponse from '@triniti/cms/plugins/pbjx/actions/clearResponse';
 import destroyEditor from '@triniti/cms/plugins/blocksmith/actions/destroyEditor';
 import NodeStatus from '@gdbots/schemas/gdbots/ncr/enums/NodeStatus';
 import startCase from 'lodash/startCase';
-import changeNodeFlow from './changeNodeFlow';
+
+import restoreFormFlow from './restoreFormFlow';
+import changeNodeFlow, { successFlow } from './changeNodeFlow';
 
 export function* onAfterSuccessFlow({ config, history, match, resolve }) {
   yield call(resolve);
@@ -16,11 +18,18 @@ export function* onAfterSuccessFlow({ config, history, match, resolve }) {
     yield put(reset(config.formName));
     yield call(history.push, match.url.match(/\/.+?\/.+?\//)[0]);
   }
+
+  if (config.shouldForceSave) {
+    yield restoreFormFlow(config);
+  }
 }
 
-export function* onAfterFailureFlow({ reject }, error) {
+export function* onAfterFailureFlow({ config, reject }, error) {
   const message = typeof error.getMessage === 'function' ? error.getMessage() : error.message;
   yield call(reject, new SubmissionError({ _error: message }));
+  if (config.shouldForceSave) {
+    config.clearFormData(config.formDataKey);
+  }
 }
 
 export function* publishAfterUpdateFlow(action) {
@@ -54,6 +63,12 @@ export default function* (action) {
   const pbj = action.pbj;
   const nodeSchema = config.schemas.node
     || config.schemas.nodes.find((node) => node.getCurie().getMessage() === match.params.type);
+
+  if (config.shouldDisableSave) {
+    yield call(successFlow, '', () => onAfterSuccessFlow(action));
+    return;
+  }
+
   yield call(changeNodeFlow, {
     expectedEvent: config.schemas.nodeUpdated.getCurie().toString(),
     failureMessage: `Save ${startCase(nodeSchema.getCurie().getMessage())} failed: `,