+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 
 ## vN.N.N
+* Create permission based force-save button for articles.
 * Adding support for `recommendations_enabled` field.
 * Use `@triniti/schemas` v1.1.12